requirements.txt
Django==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.7
